wip: aidrop actions
