+++ src/containers/main/Airdrop/mini/Actions.style.ts
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import { Typography } from '@app/components/elements/Typography.tsx';
 
 export const Wrapper = styled.div`
     flex-direction: column;
@@ -19,19 +18,9 @@ export const GemImg = styled.img`
     transform: scaleX(-1);
 `;
 
-export const ActionWrapper = styled.div`
-    flex-direction: column;
-    max-height: 60px;
-    width: 60px;
-    padding: 10px 0;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-`;
-
-export const ActionText = styled(Typography)`
-    color: ${({ theme }) => theme.palette.text.primary};
-    font-size: 11px;
-    font-weight: 600;
-    text-transform: uppercase;
+export const Avatar = styled.div`
+    width: 36px;
+    height: 36px;
+    border-radius: 50%;
+    flex-shrink: 0;
 `;
