+++ src/containers/main/Airdrop/mini/components/Action.tsx
@@ -1,0 +1,16 @@
+import { ReactNode } from 'react';
+import { ActionText, ActionWrapper } from './Action.style.ts';
+
+interface ActionProps {
+    children: ReactNode;
+    text?: string;
+}
+
+export function Action({ children, text }: ActionProps) {
+    return (
+        <ActionWrapper>
+            {children}
+            <ActionText>{text}</ActionText>
+        </ActionWrapper>
+    );
+}
